+++ src/card/Card.stories.tsx
@@ -12,15 +12,6 @@ const meta = {
   title: "Decked/Card",
   component: Card,
   tags: ["autodocs"],
-  argTypes: {
-    suit: {
-      options: ["hearts", "spades", "diamonds", "clubs"],
-      control: "select"
-    },
-    rank: {
-      control: { type: "number", min: 1, max: 13 }
-    }
-  },
   decorators: [
     (Story) => (
       <StoreProvider store={store}>
@@ -38,9 +29,6 @@ type Story = StoryObj<typeof meta>;
 // More on writing stories with args: https://storybook.js.org/docs/7.0/react/writing-stories/args
 export const AceOfSpades: Story = {
   args: {
-    id: "1",
-    suit: "spades",
-    colour: "black",
-    rank: 1
+    id: "1"
   }
 };